% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_absorption_coef.R
\name{calc_absorption_coef}
\alias{calc_absorption_coef}
\title{Convert absorbance units to Napierian or decadic absorption coefficients}
\usage{
calc_absorption_coef(absorbance, method = "napierian", path_length = 0.01)
}
\arguments{
\item{absorbance}{(num) Absorbance unit}

\item{method}{Either "napierian" or "decadic". Default method is "napierian".}

\item{path_length}{(num) Cuvette path length in meters}
}
\value{
Absorption coefficient(s) (m^-1)
}
\description{
Convert absorbance units to Napierian or decadic absorption coefficients
}
