% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_slope_ratio.R
\name{calc_slope_ratio}
\alias{calc_slope_ratio}
\title{Calculate the spectral slope ratio}
\usage{
calc_slope_ratio(wavelength, absorption, limit_of_quantification = 0)
}
\arguments{
\item{wavelength}{(num) Vector of wavelengths (nm)}

\item{absorption}{(num) Vector of absorption coefficients (m^-1)}

\item{limit_of_quantification}{(num) Instrument lower limit of quantification (m^-1) for the 275-400 nm region. Defaults to 0.}
}
\value{
(num) Ratio of spectral slopes in the 275-295 to 250-400 nm regions
}
\description{
Calculates the ratio of log-transformed slopes in the 275-295 to 250-400 nm regions
}
