% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spectral_slope.R
\name{calc_spectral_slope}
\alias{calc_spectral_slope}
\title{Calculate CDOM absorption spectral slope}
\usage{
calc_spectral_slope(
  wavelength,
  absorption,
  start = 275,
  end = 295,
  limit_of_quantification = 0
)
}
\arguments{
\item{wavelength}{(num) Vector of wavelengths (nm)}

\item{absorption}{(num) Vector of absorption coefficients (m^-1)}

\item{start}{(num) Start of the slope region (nm). Default value is 275.}

\item{end}{(num) End of the slope region (nm). Default is value is 295.}

\item{limit_of_quantification}{(num) Instrument lower limit of quantification for the specified wavelength region (m^-1). Default value is 0 (i.e. no cutoff).}
}
\value{
Log transformed spectral slope in specified region
}
\description{
This function calculates the slope of log-transformed absorption spectra. Spectral slope within the 275-295 nm region is often used as a proxy for DOM molecular weight.
}
\details{
See Helms et al. (2008) "Absorption spectral slopes and slope ratios as indicators of molecular weight, source, and photobleaching of chromophoric dissolved organic matter". https://doi.org/10.4319/lo.2008.53.3.0955
}
