% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SUVA.R
\name{calc_SUVA}
\alias{calc_SUVA}
\title{Calculate Specific Ultraviolet Absorbance (SUVA)}
\usage{
calc_SUVA(
  wavelength,
  absorption,
  DOC,
  specific_wavelength = 254,
  absorption_coef = "napierien",
  DOC_unit = "mg"
)
}
\arguments{
\item{wavelength}{(num) Vector of integer wavelengths}

\item{absorption}{(num) Vector of absorption coefficients (m^-1)}

\item{DOC}{(num) Dissolved organic carbon concentration}

\item{specific_wavelength}{(num) Wavelength (nm) to calculate SUVA at. Default is 254.}

\item{absorption_coef}{(chr) Either "napierien" or "decadic". Default is "napierien". If "napierien", will convert to absorption to decadic method for the SUVA calculation.}

\item{DOC_unit}{(chr) Either "umol" or "mg". Is 'DOC' in units micromol C per liter or milligrams carbon per liter? Default is "mg".}
}
\value{
Specific ultraviolet absorption at the specified wavelength, normalized by dissolved organic carbon concentration
}
\description{
Specific ultraviolent absorbance (SUVA) is the UV absorption of a water sample at a particular wavelength, normalized by dissolved organic carbon (DOC) concentration.
}
\details{
SUVA at 254 nm is commonly used as an indicator of DOC aromaticity (Weishaar et al. 2003). SUVA should be calculated using decadic rather than Napierien absorption coefficients and DOC concentrations should be milligrams carbon per liter.
}
\examples{


}
