% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_absorption_coef.R
\name{find_absorption_coef}
\alias{find_absorption_coef}
\title{Find the absorption at a specific wavelength}
\usage{
find_absorption_coef(wavelength, absorption, specific_wavelength)
}
\arguments{
\item{wavelength}{(num) Vector of wavelengths (nm)}

\item{absorption}{(num) Vector of absorption coefficients (m^-1)}

\item{specific_wavelength}{(num) Wavelength (nm) that you want to find CDOM absorption at}
}
\value{
(num) Absorption coefficient the specified wavelength
}
\description{
Find the absorption at a specific wavelength
}
