% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_absorbance.R
\name{interpolate_absorbance}
\alias{interpolate_absorbance}
\title{Interpolate raw absorbance values to get absorbance values at integer wavelengths}
\usage{
interpolate_absorbance(wavelength, absorbance, start = 200, end = 800)
}
\arguments{
\item{wavelength}{(num) Vector of raw wavelength values}

\item{absorbance}{(num) Vector of raw absorbance values}

\item{start}{(num) Lower limit of wavelength (nm) spectra. Default value is 200.}

\item{end}{(num) Upper limit of wavelength (nm) spectra. Default value is 800.}
}
\value{
Returns a dataframe with two columns, "wavelength" and "absorbance"
}
\description{
Some spectrophotometers return data where wavelengths are not whole numbers. This function uses linear interpolation to produce absorbance values at integer wavelengths.
}
\examples{


}
